pyproject.toml
[project]
name = "file-renamer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "tenacity>=9.1.2",
    "typer>=0.21.1",
    "uvicorn>=0.40.0",
]

[project.scripts]
renamer = "src.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src"]


[dependency-groups]
dev = [
    "ruff>=0.14.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

